`readlink -f` is not available on MacOS X
...because the canonicalisaion options to readlink, including -f, are GNU
coreutils specific.

This patch implements a more portable (but not as strictly correct) way of
finding the directory in which the script resides. If the called script
itself is a symlink then we now get the canonical name of the directory in
which the symlink resides, rather than the canonical name of the directory
in which the target of the symlink resides but I think we can live with
that for now.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>