A quick and dirty test suite
We build the Blue Badge example in a fresh directory and use it to
ensure that the .proto files that result from the main build process
seem to work.

I'm not convinced that it catches everything because Simon was
suggesting that he has been getting some warnings from the Protocol
Buffers Compiler that he uses from libprotoc-3.3.0. However, it detects
the case when fields are renamed in the specification and not in the
examples.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>